# backend/src/internal/server/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "✨Reverse-engineered API for Gemini web app. It can be used as a genuine API key from OpenAI, Gemini, and Claude.",
        "title": "Gemini Web To API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4981",
    "basePath": "/",
    "paths": {
...
        "/openai/v1/chat/completions": {
            "post": {
                "description": "Generates a completion for the chat message. Supports both standard JSON and streaming (SSE) response.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "OpenAI"
                ],
                "summary": "Chat Completions (OpenAI)",
                "parameters": [
                    {
                        "description": "Chat Completion Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gemini-web-to-api_internal_modules_openai_dto.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of dto.ChatCompletionChunk JSON objects",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gemini-web-to-api_internal_modules_openai_dto.ChatCompletionMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_openai_dto.ChatCompletionRequest": {
            "type": "object",
            "properties": {
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_openai_dto.ChatCompletionMessage"
                    }
                },
                "model": {
                    "type": "string"
                },
                "stream": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "tool_choice": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-web-to-api_internal_modules_openai_dto.ToolDefinition"
                    }
                }
            }
        },
...
        "gemini-web-to-api_internal_modules_openai_dto.ToolDefinition": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/gemini-web-to-api_internal_modules_openai_dto.ToolFunctionDefinition"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "gemini-web-to-api_internal_modules_openai_dto.ToolFunctionDefinition": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}